pyproject.toml
[project]
name = "text2sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clickhouse-connect>=0.10.0",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.1",
    "langchain-google-genai>=3.2.0",
    "langgraph>=1.0.4",
    "langgraph-cli[inmem]>=0.4.7",
    "mysql-connector-python>=9.5.0",
    "pandas>=2.3.3",
    "paramiko>=4.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.16.1",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "sshtunnel>=0.4.0",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]
